// customer-app/src/types/booking.ts
export type EventLocation = {
  placeId: string | null;
  formattedAddress: string;
  latitude: number | null;
  longitude: number | null;
  city: string;
  district: string;
  state: string;
  pincode: string;
  manuallyEdited: boolean;
};